Moved file format realted classes into FileFormat package
