+++ pulumi_snowflake/__init__.py
@@ -1,11 +1,6 @@
-from .file_format import FileFormat
-from .file_format_provider import FileFormatProvider
-from .file_format_type import FileFormatType
 from .snowflake_connection_provider import SnowflakeConnectionProvider
 
 __all__ = [
-    FileFormat,
-    FileFormatProvider,
-    SnowflakeConnectionProvider,
-    FileFormatType
+    'fileformat',
+    "SnowflakeConnectionProvider"
 ]
