+++ pulumi_snowflake/fileformat/file_format_provider.py
@@ -1,6 +1,6 @@
 from pulumi.dynamic import CreateResult, DiffResult, ResourceProvider
 from pulumi_snowflake.random_id import RandomId
-from pulumi_snowflake.snowflake_connection_provider import SnowflakeConnectionProvider
+from pulumi_snowflake import SnowflakeConnectionProvider
 from pulumi_snowflake.validation import Validation
