+++ test/file_format_provider_tests_delete.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import Mock, call
 
-from pulumi_snowflake.file_format_provider import FileFormatProvider
+from pulumi_snowflake.fileformat import FileFormatProvider
 
 
 class FileFormatProviderTests(unittest.TestCase):
